# docs/schemas/books-api-v3.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Books client API",
    "version": "1.2.0",
    "description": "Experimental statement import API. Additive contract snapshot; routes and envelopes remain v1. Read grants permit match inspection; uploads and previews require import; posting requires post. See ../statement-formats.md for supported profiles, source options, identity review, limits, and single-currency accounting. *_cents and int64 fields are integer strings. Each entity has one supported monetary currency; no currency conversion. Fields named *_cents are integer minor-unit strings in the owning currency, not necessarily hundredths. Source amounts remain decimal strings. Discover supported currencies through /v1/capabilities."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8484"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "summary": "Capabilities",
        "operationId": "get_v1_capabilities",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schema": {
                      "type": "string",
                      "const": "books.api/v1"
                    },
                    "ok": {
                      "const": true
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "schema",
                    "ok",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "See docs/api.md for domain codes and HTTP error statuses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/companies": {
      "get": {
        "summary": "List granted companies",
        "operationId": "get_v1_companies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schema": {
                      "type": "string",
                      "const": "books.api/v1"
                    },
                    "ok": {
                      "const": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Company"
                      }
                    }
                  },
                  "required": [
                    "schema",
                    "ok",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "See docs/api.md for domain codes and HTTP error statuses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "schema": {
            "type": "string",
            "const": "books.api/v1"
          },
          "ok": {
            "const": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "schema",
          "ok",
          "error"
        ]
      },
...
      "Company": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "entity": {
            "type": "string"
          },
          "book": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "basis": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "name",
          "entity",
          "book",
          "currency",
          "basis"
        ]
      },
...
    }
  }
}